requirements.txt
Django==5.2.1
python-dotenv==1.1.0
psycopg2==2.9.10
djangorestframework==3.16.0
drf-yasg==1.21.10
celery==5.5.2
redis==6.1.0
mock==5.2.0